requirements.txt
requests==2.31.0 
pandas==2.2.2 
scikit-learn==1.5.1 
solders==0.21.0 
python-telegram-bot==20.7 
flask==3.0.3 
reportlab==4.2.2 
